feat(openaev): introducing processors for proper code architecture (#317)
