+++ openaev/openaev/processors/__init__.py
@@ -1,0 +1,2 @@
+from openaev.processors.json_api_processor import JsonApiProcessor
+from openaev.processors.json_flat_processor import JsonFlatProcessor
